output: Move printers into own dir
