+++ coalib/collecting/BearCollector.py
@@ -6,7 +6,7 @@ import sys
 
 from coalib.collecting.FileCollector import FileCollector
 from coalib.misc.StringConstants import StringConstants
-from coalib.output.ConsolePrinter import ConsolePrinter
+from coalib.output.printers.ConsolePrinter import ConsolePrinter
 from coalib.settings.Section import Section
 from coalib.settings.Setting import path_list
