+++ coalib/output/Interactor.py
@@ -1,6 +1,6 @@
 from coalib.bearlib.abstractions.SectionCreatable import SectionCreatable
 from coalib.bears.results.Result import Result
-from coalib.output.ConsolePrinter import ConsolePrinter
+from coalib.output.printers.ConsolePrinter import ConsolePrinter
 from coalib.misc.i18n import _
 from coalib.settings.FunctionMetadata import FunctionMetadata
