+++ coalib/output/printers/ConsolePrinter.py
@@ -1,8 +1,8 @@
 import sys
 
-from coalib.output.ColorPrinter import ColorPrinter
-from coalib.output.LOG_LEVEL import LOG_LEVEL
-from coalib.output.LogPrinter import LogPrinter
+from coalib.output.printers.ColorPrinter import ColorPrinter
+from coalib.output.printers.LogPrinter import LogPrinter
+from coalib.output.printers.LOG_LEVEL import LOG_LEVEL
 
 
 class ConsolePrinter(LogPrinter, ColorPrinter):
