+++ coalib/tests/processes/communication/LogMessageTest.py
@@ -2,7 +2,7 @@ import sys
 
 sys.path.insert(0, ".")
 from coalib.misc.i18n import _
-from coalib.output.LOG_LEVEL import LOG_LEVEL
+from coalib.output.printers.LOG_LEVEL import LOG_LEVEL
 from coalib.misc.StringConstants import StringConstants
 from coalib.processes.communication.LogMessage import LogMessage
 import unittest
